pyproject.toml
[project]
name = "vibelens"
version = "0.9.30"
description = "Agent Trajectory visualization and analysis platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "VibeLens at ChatsLab", email = "chatslab001@gmail.com" },
]
keywords = ["agent", "trajectory", "analysis", "claude", "codex", "gemini", "visualization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "typer>=0.15.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.22",
    "litellm>=1.40.0",
    "jinja2>=3.1.0",
    "tiktoken>=0.7.0",
    "httpx>=0.27.0",
    "cachetools>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
vibelens = "vibelens.cli:app"

[project.urls]
Homepage = "https://vibelens.chats-lab.org/"
Repository = "https://github.com/yejh123/VibeLens"
"Bug Tracker" = "https://github.com/yejh123/VibeLens/issues"
Changelog = "https://github.com/yejh123/VibeLens/releases"
Demo = "https://vibelens.chats-lab.org/"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vibelens"]

[tool.ruff]
src = ["src"]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["vibelens"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests that make real LLM calls (deselect with '-m \"not slow\"')",
]
addopts = "-m 'not slow'"
